use lws_plat_file_* instead of compatible_file_* fix linux and mbed
